refactor: build.gradle 1차 정리 및 test 실행 파일 수정
