+++ domain-kafka/src/test/java/com/kernelsquare/domainkafka/DomainKafkaApplicationTests.java
@@ -1,8 +1,10 @@
 package com.kernelsquare.domainkafka;
 
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
+@Disabled
 @SpringBootTest
 class DomainKafkaApplicationTests {
