+++ domain-mongodb/src/test/java/com/kernelsquare/domainmongodb/DomainMongoApplicationTests.java
@@ -1,9 +1,7 @@
-package com.kernelsquare.domainmongodb.domain;
+package com.kernelsquare.domainmongodb;
 
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
 public class DomainMongoApplicationTests {
-    public void contextLoads() {
-    }
 }
