+++ domain-redis/src/test/java/com/kernelsquare/domainredis/DomainRedisApplicationTests.java
@@ -1,8 +1,10 @@
 package com.kernelsquare.domainredis;
 
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
+@Disabled
 @SpringBootTest
 class DomainRedisApplicationTests {
